11/10

+++ 作业与课堂练习/CSUOJ/石子游戏.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-	long int n,ai,i;
-	while (scanf("%ld",&n)!=EOF){
+	int n,ai,i;
+	while (scanf("%d",&n)!=EOF){
 		for (i=0;i<n;i++) scanf("%d",&ai);
 		if (n>1) printf("NO\n");
 		else if (ai%2) printf("NO\n");
